cpu: aarch64: make softmax ops use stateless ACL interface
